lmb-796: create a generic method to check if the ids are available
